Remove redundant PoliteiaFragment
