+++ app/src/main/java/com/dcrandroid/fragments/BaseFragment.kt
@@ -9,7 +9,6 @@ package com.dcrandroid.fragments
 import androidx.annotation.StringRes
 import androidx.fragment.app.Fragment
 import com.dcrandroid.HomeActivity
-import com.dcrandroid.activities.more.PoliteiaActivity
 import com.dcrandroid.data.Transaction
 import com.dcrandroid.util.WalletData
 import com.google.gson.Gson
@@ -64,13 +63,6 @@ open class BaseFragment : Fragment(), SyncProgressListener, TxAndBlockNotificati
     fun setToolbarTitle(@StringRes title: Int, showShadow: Boolean) {
         if (context != null) {
             setToolbarTitle(context!!.getString(title), showShadow)
-        }
-    }
-
-    fun syncProposals() {
-        if (activity is PoliteiaActivity) {
-            val politeiaActivity = activity as PoliteiaActivity
-            politeiaActivity.syncProposals()
         }
     }
