--- app/src/main/res/layout/fragment_politeia.xml
@@ -1,186 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ Copyright (c) 2018-2019 The Decred developers
-  ~ Use of this source code is governed by an ISC
-  ~ license that can be found in the LICENSE file.
-  -->
-
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:orientation="vertical"
-    android:background="@color/colorBackground"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <com.google.android.material.appbar.AppBarLayout
-        android:layout_width="match_parent"
-        android:layout_height="?actionBarSize"
-        android:id="@+id/app_bar"
-        app:elevation="0dp"
-        android:background="@color/colorBackground"
-        android:theme="@style/AppTheme.AppBarOverlay">
-
-        <androidx.appcompat.widget.Toolbar
-            android:id="@+id/toolbar"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:paddingStart="0dp"
-            android:paddingLeft="0dp"
-            app:contentInsetStart="0dp"
-            app:contentInsetLeft="0dp"
-            android:paddingEnd="0dp"
-            android:paddingRight="0dp"
-            app:contentInsetEnd="0dp"
-            app:contentInsetRight="0dp">
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:paddingStart="@dimen/margin_padding_size_8"
-                android:paddingEnd="0dp"
-                android:orientation="horizontal"
-                android:gravity="center_vertical">
-
-                <ImageView
-                    android:layout_width="40dp"
-                    android:layout_height="40dp"
-                    android:padding="@dimen/margin_padding_size_8"
-                    android:background="@drawable/bg_primary_ripple"
-                    android:focusable="true"
-                    android:clickable="true"
-                    android:id="@+id/go_back"
-                    app:srcCompat="@drawable/ic_back" />
-
-                <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:id="@+id/app_bar_title"
-                    android:text="@string/politeia"
-                    android:layout_marginStart="@dimen/margin_padding_size_8"
-                    android:textSize="@dimen/edit_text_size_20"
-                    android:textColor="@color/darkBlueTextColor"
-                    android:includeFontPadding="false"
-                    app:fontFamily="@font/source_sans_pro" />
-
-            </LinearLayout>
-
-        </androidx.appcompat.widget.Toolbar>
-
-    </com.google.android.material.appbar.AppBarLayout>
-
-    <LinearLayout
-        android:orientation="horizontal"
-        android:background="@color/colorBackground"
-        android:paddingLeft="@dimen/margin_padding_size_8"
-        android:paddingRight="@dimen/margin_padding_size_8"
-        android:paddingBottom="@dimen/margin_padding_size_12"
-        android:id="@+id/proposals_page_header"
-        android:elevation="0dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <Spinner
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:id="@+id/timestamp_sort_spinner" />
-
-        <Spinner
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:id="@+id/category_sort_spinner" />
-
-        <LinearLayout
-            android:id="@+id/sync_layout"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:background="@drawable/reconnect_background"
-            android:clickable="true"
-            android:focusable="true"
-            android:gravity="center_vertical"
-            android:orientation="horizontal"
-            android:paddingStart="@dimen/margin_padding_size_4"
-            android:paddingTop="@dimen/margin_padding_size_4"
-            android:paddingEnd="9dp"
-            android:paddingBottom="@dimen/margin_padding_size_4">
-
-            <ImageView
-                android:layout_width="16dp"
-                android:layout_height="16dp"
-                app:srcCompat="@drawable/ic_rescan" />
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="@dimen/margin_padding_size_4"
-                android:includeFontPadding="false"
-                android:text="@string/sync"
-                android:textColor="@color/darkBlueTextColor"
-                android:textSize="@dimen/edit_text_size_14"
-                android:id="@+id/tv_reconnect"
-                app:fontFamily="@font/source_sans_pro" />
-
-        </LinearLayout>
-
-    </LinearLayout>
-
-    <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
-        android:id="@+id/swipe_refresh_layout"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:visibility="visible">
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/recycler_view"
-            android:background="@drawable/card_bg"
-            android:elevation="@dimen/margin_padding_size_4"
-            android:layout_marginTop="@dimen/margin_padding_size_4"
-            android:layout_marginStart="@dimen/margin_padding_size_8"
-            android:layout_marginEnd="@dimen/margin_padding_size_8"
-            android:layout_marginBottom="@dimen/margin_padding_size_8" />
-
-    </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
-
-    <TextView
-        android:id="@+id/empty_list"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:gravity="center"
-        android:textSize="@dimen/edit_text_size_16"
-        android:layout_marginBottom="@dimen/margin_padding_size_18"
-        android:textColor="@color/lightGrayTextColor"
-        app:fontFamily="@font/source_sans_pro"
-        android:text="@string/no_proposals"/>
-
-    <LinearLayout
-        android:id="@+id/loading_view"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:gravity="center"
-        android:orientation="vertical"
-        android:visibility="gone">
-
-        <ProgressBar
-            android:id="@+id/progressBar"
-            style="?android:attr/progressBarStyleHorizontal"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:indeterminate="false"
-            android:max="100"
-            android:minWidth="200dp"
-            android:minHeight="50dp"
-            android:progress="1" />
-
-        <TextView
-            android:id="@+id/textView"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:gravity="center"
-            android:layout_marginBottom="@dimen/margin_padding_size_32"
-            />
-
-    </LinearLayout>
-
-</LinearLayout>
